utils: cast

+++ src/utils.rs
@@ -123,7 +123,7 @@ pub fn name_to_sid(name: &str, system: Option<&str>) -> Result<Vec<BYTE>, DWORD>
         return Err(unsafe { GetLastError() });
     }
 
-    unsafe { sid.set_len(sid_size) };
+    unsafe { sid.set_len(sid_size as usize) };
 
     Ok(sid)
 }
